requirement.txt
numpy==1.17.2
transforms3d==0.3.1
scipy==1.3.1
matplotlib==3.1.1
yacs==0.1.6
tensorflow==2.0.0
emoji
pillow==6.2.1